Total Re-Design
Moved nodes and tracker out of the same file.
Generilzed the AStar class so you can run seperate paths at once without
mixing anything up.
Placing a building places a wall onto the map. All things that will
utilize A* need to have the Tracker class or copy some of its functions.
The tracker class is more of an example of how to use the AStar class.