More UI cleanup
There are still some scripts that i'm unsure about... Currently unused but might be needed for the Demo.
Need to check keys used for certain things: "z" is bound to rotating buildings and camera zoom. Should probably create a script that stores all controls, this will make it easier to change controls as well.